tools/requirements.txt
altgraph==0.16.1
cycler==0.10.0
kiwisolver==1.1.0
macholib==1.11
matplotlib==3.1.1
numpy==1.17.0
psutil==5.6.3
PyInstaller==3.5
pyparsing==2.4.2
python-dateutil==2.8.0
PyYAML==5.1.2
six==1.12.0
